introduce abstract restart_tests class
